chore: refactored user
