+++ .github/workflows/security-audit.yml
@@ -4,7 +4,6 @@ on:
   push:
     branches:
       - main
-      - development
     paths:
       - "**/Cargo.toml"
       - "**/Cargo.lock"
