+++ backend/src/users/mod.rs
@@ -1,0 +1,5 @@
+pub mod auth;
+pub mod model;
+
+pub use auth::*;
+pub use model::*;
